🐛  fix storage for web

+++ client/lib/supabase.ts
@@ -1,16 +1,29 @@
 import 'react-native-url-polyfill/auto'
 import * as SecureStore from 'expo-secure-store';
+import AsyncStorage from '@react-native-async-storage/async-storage';
 import { createClient } from '@supabase/supabase-js'
+import { Platform } from 'react-native';
 
 const ExpoSecureStoreAdapter = {
   getItem: (key: string) => {
-    return SecureStore.getItemAsync(key);
+    if (Platform.OS === 'web') {
+      return AsyncStorage.getItem(key)
+    }
+    else {
+      return SecureStore.getItemAsync(key);
+    }
   },
   setItem: (key: string, value: string) => {
-    SecureStore.setItemAsync(key, value);
+    if (Platform.OS === 'web') {
+      AsyncStorage.setItem(key, value)
+    }
+    else { SecureStore.setItemAsync(key, value); }
   },
   removeItem: (key: string) => {
-    SecureStore.deleteItemAsync(key);
+    if (Platform.OS === 'web') { AsyncStorage.removeItem(key) }
+    else {
+      SecureStore.deleteItemAsync(key);
+    }
   },
 };
